Refactor the Potcar class to the potential module
Ultimately to avoid recursive imports and still follow pep8 imports at head.
